third_party/WebKit/LayoutTests/inspector/elements/styles/undo-add-new-rule-expected.txt

Tests that adding a new rule can be undone.

Text
After adding new rule:
[expanded] 
element.style  { ()
font-size: 12px;

[expanded] 
div.foo  { (inspector-stylesheet:3 -> inspector-stylesheet:3:1)

[expanded] 
div.foo  { (inspector-stylesheet:1 -> inspector-stylesheet:1:1)

[expanded] 
div  { (user agent stylesheet)
display: block;


===== Style sheet text: =====
div.foo {}

div.foo {}
=============================
After undo:
[expanded] 
element.style  { ()
color: red;

[expanded] 
div.foo  { (inspector-stylesheet:1 -> inspector-stylesheet:1:1)

[expanded] 
div  { (user agent stylesheet)
display: block;


===== Style sheet text: =====
div.foo {}
=============================
After redo:
[expanded] 
element.style  { ()
font-size: 12px;

[expanded] 
div.foo  { (inspector-stylesheet:3 -> inspector-stylesheet:3:1)

[expanded] 
div.foo  { (inspector-stylesheet:1 -> inspector-stylesheet:1:1)

[expanded] 
div  { (user agent stylesheet)
display: block;


===== Style sheet text: =====
div.foo {}

div.foo {}
=============================
